fixed test failure
