+++ config/sawyer.yaml
@@ -5,14 +5,14 @@ dh_parameter:
   joint4: [0, -0.1685, 0, -1.57079633]
   joint5: [0, 0.4, 0, 1.57079633]
   joint6: [0, 0.1363, 0, -1.57079633]
-  joint7: [1.57079633, 0.13375, 0, 1.57079633]
+  joint7: [3.14159265, 0.13375, 0, 1.57079633]
 
 su_dh_parameter:
   su1: [1.57079633, -0.157, -1.57079633, 0.07, 0, 1.57079633]
   su2: [-1.57079633, -0.0925, 1.57079633, 0.07, 0, 1.57079633]
   su3: [-1.57079633, -0.16, 1.57079633, 0.05, 0, 1.57079633]
   su4: [-1.57079633, 0.0165, 1.57079633, 0.05, 0, 1.57079633]
-  su5: [-1.57079633, -0.17, 1.57079633, 0.04, 0, 1.57079633]
+  su5: [-1.57079633, -0.17, 1.57079633, 0.05, 0, 1.57079633]
   su6: [-1.57079633, 0.0053, 1.57079633, 0.04, 0, 1.57079633]
   su7: [0.0, 0.12375, 0.0, 0.03, 0, -1.57079633]
